+++ CSharpExcelChangeLoggerTest/ChangeLogger/Memory/ChangeHandlerMemoryTest.cs
@@ -1,4 +1,4 @@
-﻿using CSharpExcelChangeLogger.ChangeLogger.Memory;
+using CSharpExcelChangeLogger.ChangeLogger.Memory;
 using CSharpExcelChangeLogger.Excel;
 using CSharpExcelChangeLoggerTest.Mock;
 using NUnit.Framework;
@@ -17,7 +17,7 @@ namespace CSharpExcelChangeLoggerTest.ChangeLogger.Memory
             IWorksheet sheet = new SimpleMockSheet();
             IRange range = new SimpleMockRange();
 
-            IMemoryComparison comparison = memory.DoesMemoryMatch(sheet, range);
+            IMemoryComparison comparison = memory.Compare(sheet, range);
             Assert.AreEqual(false, comparison.LocationMatches, "Location should not match given sheet/range if no data saved to memory");
             Assert.AreEqual(false, comparison.LocationMatchesAndDataMatches, 
                 "Location and data should not match given sheet/range if no data saved to memory");
@@ -31,7 +31,7 @@ namespace CSharpExcelChangeLoggerTest.ChangeLogger.Memory
             IRange range = new SimpleMockRange();
 
             memory.SetMemory(sheet, range);
-            IMemoryComparison comparison = memory.DoesMemoryMatch(sheet, range);
+            IMemoryComparison comparison = memory.Compare(sheet, range);
 
             Assert.AreEqual(true, comparison.LocationMatches, "Location should match given sheet/range");
             Assert.AreEqual(true, comparison.LocationMatchesAndDataMatches,
@@ -47,7 +47,7 @@ namespace CSharpExcelChangeLoggerTest.ChangeLogger.Memory
             IRange range2 = new SimpleMockRange("2");
 
             memory.SetMemory(sheet, range1);
-            IMemoryComparison comparison = memory.DoesMemoryMatch(sheet, range2);
+            IMemoryComparison comparison = memory.Compare(sheet, range2);
 
             Assert.AreEqual(false, comparison.LocationMatches, "Location should not match because the saved range has a different address");
             Assert.AreEqual(false, comparison.LocationMatchesAndDataMatches,
@@ -65,7 +65,7 @@ namespace CSharpExcelChangeLoggerTest.ChangeLogger.Memory
             range2.RangeData = new string[2, 2] { { "1", "2" }, { "3", "4" } };
 
             memory.SetMemory(sheet, range1);
-            IMemoryComparison comparison = memory.DoesMemoryMatch(sheet, range2);
+            IMemoryComparison comparison = memory.Compare(sheet, range2);
 
             Assert.AreEqual(true, comparison.LocationMatches, "Location should match because the saved range has the same address");
             Assert.AreEqual(false, comparison.LocationMatchesAndDataMatches,
@@ -83,7 +83,7 @@ namespace CSharpExcelChangeLoggerTest.ChangeLogger.Memory
             range2.RangeData = new string[2, 2] { { "one", "two" }, { "three", "four" } };
 
             memory.SetMemory(sheet, range1);
-            IMemoryComparison comparison = memory.DoesMemoryMatch(sheet, range2);
+            IMemoryComparison comparison = memory.Compare(sheet, range2);
 
             Assert.AreEqual(true, comparison.LocationMatches, "Location should match because the saved range has the same address");
             Assert.AreEqual(true, comparison.LocationMatchesAndDataMatches,
@@ -102,7 +102,7 @@ namespace CSharpExcelChangeLoggerTest.ChangeLogger.Memory
             range.RangeData = new string[2, 2] { { "one", "two" }, { "three", "four" } };
 
             memory.SetMemory(sheet, range);
-            IMemoryComparison comparison = memory.DoesMemoryMatch(sheet, range);
+            IMemoryComparison comparison = memory.Compare(sheet, range);
 
             Assert.AreEqual(true, comparison.LocationMatches, "Location should match because the saved range has the same address");
             Assert.AreEqual(false, comparison.LocationMatchesAndDataMatches,
@@ -124,7 +124,7 @@ namespace CSharpExcelChangeLoggerTest.ChangeLogger.Memory
 
             memory.SetMemory(sheet1, range);
             range = new RowChangeMockRange();
-            IMemoryComparison comparison = memory.DoesMemoryMatch(sheet2, range);
+            IMemoryComparison comparison = memory.Compare(sheet2, range);
 
             Assert.AreEqual(true, comparison.IsNewRow, "Should report new row as sheet row count increased");
             Assert.AreEqual(false, comparison.IsNewColumn, "Should not report new column as sheet column count remained the same");
@@ -147,7 +147,7 @@ namespace CSharpExcelChangeLoggerTest.ChangeLogger.Memory
 
             memory.SetMemory(sheet1, range);
             range = new ColumnChangeMockRange();
-            IMemoryComparison comparison = memory.DoesMemoryMatch(sheet2, range);
+            IMemoryComparison comparison = memory.Compare(sheet2, range);
 
             Assert.AreEqual(false, comparison.IsNewRow, "Should not report new row as sheet row count remained the same");
             Assert.AreEqual(true, comparison.IsNewColumn, "Should report new column as sheet row cound increased");
@@ -172,7 +172,7 @@ namespace CSharpExcelChangeLoggerTest.ChangeLogger.Memory
             memory.SetMemory(sheet1, range);
 
             range = new RowChangeMockRange();
-            IMemoryComparison comparison = memory.DoesMemoryMatch(sheet2, range);
+            IMemoryComparison comparison = memory.Compare(sheet2, range);
 
             Assert.IsNull(comparison.DataBeforeChange, "GIVEN: Data from before change should not be saved");
             Assert.AreEqual(true, comparison.IsNewRow, "Should report new row as sheet row count increased");
@@ -198,7 +198,7 @@ namespace CSharpExcelChangeLoggerTest.ChangeLogger.Memory
             memory.SetMemory(sheet1, range);
 
             range = new ColumnChangeMockRange();
-            IMemoryComparison comparison = memory.DoesMemoryMatch(sheet2, range);
+            IMemoryComparison comparison = memory.Compare(sheet2, range);
 
             Assert.IsNull(comparison.DataBeforeChange, "GIVEN: Data from before change should not be saved");
             Assert.AreEqual(false, comparison.IsNewRow, "Should not report new row as sheet row count remained the same");
@@ -222,7 +222,7 @@ namespace CSharpExcelChangeLoggerTest.ChangeLogger.Memory
 
             memory.SetMemory(sheet1, range);
             range = new RowChangeMockRange();
-            IMemoryComparison comparison = memory.DoesMemoryMatch(sheet2, range);
+            IMemoryComparison comparison = memory.Compare(sheet2, range);
 
             Assert.AreEqual(false, comparison.IsNewRow, "Should not report new row as sheet row count decreased");
             Assert.AreEqual(false, comparison.IsNewColumn, "Should not report new column as sheet column count remained the same");
@@ -246,7 +246,7 @@ namespace CSharpExcelChangeLoggerTest.ChangeLogger.Memory
             memory.SetMemory(sheet1, range);
 
             range = new ColumnChangeMockRange();
-            IMemoryComparison comparison = memory.DoesMemoryMatch(sheet2, range);
+            IMemoryComparison comparison = memory.Compare(sheet2, range);
 
             Assert.AreEqual(false, comparison.IsNewRow, "Should not report new row as sheet row count remained the same");
             Assert.AreEqual(false, comparison.IsNewColumn, "Should not report new column as sheet column count decreased");
@@ -271,7 +271,7 @@ namespace CSharpExcelChangeLoggerTest.ChangeLogger.Memory
             memory.SetMemory(sheet1, range);
 
             range = new RowChangeMockRange();
-            IMemoryComparison comparison = memory.DoesMemoryMatch(sheet2, range);
+            IMemoryComparison comparison = memory.Compare(sheet2, range);
 
             Assert.IsNull(comparison.DataBeforeChange, "GIVEN: Data from before change should not be saved");
             Assert.AreEqual(false, comparison.IsNewRow, "Should not report new row as sheet row count decreased");
@@ -295,9 +295,9 @@ namespace CSharpExcelChangeLoggerTest.ChangeLogger.Memory
 
             range.RangeData = new string[memory.MaxRangeSizeForStoringData + 1, 1];
             memory.SetMemory(sheet1, range);
-
+            
             range = new ColumnChangeMockRange();
-            IMemoryComparison comparison = memory.DoesMemoryMatch(sheet2, range);
+            IMemoryComparison comparison = memory.Compare(sheet2, range);
 
             Assert.IsNull(comparison.DataBeforeChange, "GIVEN: Data from before change should not be saved");
             Assert.AreEqual(false, comparison.IsNewRow, "Should not report new row as sheet column count remained the same");
@@ -323,7 +323,7 @@ namespace CSharpExcelChangeLoggerTest.ChangeLogger.Memory
             memory.SetMemory(sheet1, range1);
 
             SimpleMockRange range2 = new SimpleMockRange(range2Addr);
-            IMemoryComparison comparison = memory.DoesMemoryMatch(sheet2, range2);
+            IMemoryComparison comparison = memory.Compare(sheet2, range2);
 
             Assert.AreEqual(sheet1Name, comparison.SheetNameBeforeChange, "Sheet name from before change should match name of sheet1");
             Assert.AreEqual(sheet2Name, comparison.SheetNameAfterChange, "Sheet name from after change should match name of sheet2");