+++ CHANGELOG.md
@@ -7,11 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Rename DoesMemoryMatch method to Compare (#22)
+- Remove DataMatches property from IMemoryComparison interface (#22)
+- Rename SetLogger method to SetApplicationLogger to help avoid confusion (#22)
+- Sheet and range data is now cached to avoid expesive read operations on the provided sheet and range classes (#26)
+
 ### Added
 - Add properties on the memory comparison object for the range address and sheet name before and after a change (#24)
-
-### Changed
-- Sheet and range data is now cached to avoid expesive read operations on the provided sheet and range classes (#26)
 
 ## [0.1.0] - 2020-05-18